requirements.txt
# PDF Parsing
pypdf

# Embeddings (using Sentence Transformers)
sentence-transformers
torch

# Vector Store (ChromaDB - local persistence)
chromadb

# LLM (using CTransformers for GGUF models & Transformers for general utilities)
ctransformers[cuda] # Use [cuda] if you have an NVIDIA GPU, otherwise just 'ctransformers'
transformers
accelerate
llama-cpp-python